Initialize Model Associations Once & Add Testing Setup
*Before:* Model Associations were defined in a function
which was called in `app.js` to initialize them.

*Issue:* This was currently working, as during normal execution,
`models()` was only called that one time. However, in testing,
we need to reference `app` multiple times, so the Model Assocations
were being redefined in each test. This was causing tests to break,
as Sequelize was complaining that in `models/index.js` `addScope`
was adding duplicate scopes to Models (since it was being called
multiple times).

*Now:* We remove the function wrapping the Model Associations
so that the first file which imports them defines them once,
and subsequent times that file is loaded, Model Associations
will not be redefined.

Old Usage:
```js
import models from './models';
models();
```

New Usage:
```js
import './models';
```