Add Migration Adding ON DELETE ON UPDATE Constraints for quotes_tags
We want references between quotes and tags to be
deleted if quotes or tags are deleted. It doesn't
make sense to have a quote reference a tag that
doesn't exist or a tag to reference a quote
that doesn't exist.

Also without this constraint, there were issues
in testing removing the rows from these tables.

Had to handle quirks between SQLite and Postgres
for the migration. See the commit for details.

May be able to rewrite the migration when we
upgrade to a newer version of Sequelize which
has support in the queryInterface for modifying constraints.